backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# ML and Data Science
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
shap==0.43.0
xgboost==2.0.2

# PDF Generation
reportlab==4.0.7
PyPDF2==3.0.1

# Utilities
httpx==0.25.2
aiofiles==23.2.1
email-validator==2.1.0
